test: run root tests against local fhir, increase memory of fhir server to 2gb
