+++ src/helpers/fhirFunctions/translateCode.ts
@@ -6,13 +6,15 @@ import { getLogger } from '../logger';
 
 export const translateCode = async (input: string, tableId: string) => {
   const { tables } = getCache();
-  // fork: os
   try {
     let map = tables.get(tableId);
     if (map === undefined) {
       getLogger().info(`Table ${tableId} not cached, trying to fetch from server...`);
-      map = (await getTable(tableId))[tableId];
-      tables.set(tableId, map);
+      const table = await getTable(tableId);
+      if (table) {
+        map = table[tableId];
+        tables.set(tableId, map);
+      }
     };
 
     const mapFiltered = map[input];
