+++ tests/root/case1.test.ts
@@ -1,0 +1,21 @@
+import { test } from '@jest/globals';
+import request from 'supertest';
+
+import { mockInput } from '../utils/mockInput';
+
+test('Case 1 - Boolean false disappear from FLASH outputs', async () => {
+  const mapping = `
+            InstanceOf: Patient
+            * active = false
+        `;
+  const requestBody = {
+    input: mockInput,
+    fume: mapping
+  };
+
+  const res = await request(globalThis.app).post('/').send(requestBody);
+  expect(res.body).toStrictEqual({
+    resourceType: 'Patient',
+    active: false
+  });
+});
