+++ tests/utils/mockInput.ts
@@ -1,0 +1,35 @@
+
+export const mockInput = {
+  mrn: 'PP875023983',
+  status: 'active',
+  ssn: '123-45-6789',
+  passport_number: '7429184766',
+  passport_country: 'USA',
+  first_name: 'Jessica',
+  last_name: 'Rabbit',
+  birth_date: '1988-06-22',
+  sex: 'F',
+  address: {
+    city_name: 'Orlando',
+    state: 'FL',
+    street_name: 'Buena Vista',
+    house_number: 1375,
+    zip_code: '3456701',
+    lat: 28.3519592,
+    long: -81.417283
+  },
+  phones: [
+    {
+      type: 'HOME',
+      number: '+1 (407) 8372859'
+    },
+    {
+      type: 'CELL',
+      number: '+1 (305) 9831195'
+    }
+  ],
+  primary_doctor: {
+    full_name: 'Dr. Dolittle',
+    license: '1-820958'
+  }
+};
